back/requirements.txt
numpy
fastapi
uvicorn
pillow
mathplotlib
pandas
python-multipart
tensorflow 2.13.0